+++ lib/Materials/materialinventory.dart
@@ -229,25 +229,7 @@ class _MaterialInventoryState extends State<MaterialInventory> {
                                             onTap: () async {
                                               final documentRef =
                                                   snapshot.reference;
-                                              Query query = CollectionRef.where(
-                                                  'Images',
-                                                  isEqualTo: ((snapshot.data()
-                                                      as Map<String,
-                                                          dynamic>)['Images']));
-                                              query.get().then((QuerySnapshot
-                                                  snapshot) async {
-                                                if (snapshot.docs.isNotEmpty) {
-                                                  for (DocumentSnapshot document
-                                                      in snapshot.docs) {
-                                                    final documentId =
-                                                        document.reference;
-                                                    await documentId.delete();
-                                                  }
-                                                } else {
-                                                  print(
-                                                      'No documents found with the matching image URL.');
-                                                }
-                                              });
+
                                               for (int i = 0;
                                                   i <
                                                       (snapshot.data() as Map<
@@ -262,7 +244,34 @@ class _MaterialInventoryState extends State<MaterialInventory> {
                                                             as Map<String,
                                                                 dynamic>)[
                                                         'Images'][i])
-                                                    .delete();
+                                                    .delete()
+                                                    .then((value) {
+                                                  Query query =
+                                                      CollectionRef.where(
+                                                          'Images',
+                                                          isEqualTo: ((snapshot
+                                                                      .data()
+                                                                  as Map<String,
+                                                                      dynamic>)[
+                                                              'Images']));
+                                                  query.get().then(
+                                                      (QuerySnapshot
+                                                          snapshot) async {
+                                                    if (snapshot
+                                                        .docs.isNotEmpty) {
+                                                      for (DocumentSnapshot document
+                                                          in snapshot.docs) {
+                                                        final documentId =
+                                                            document.reference;
+                                                        await documentId
+                                                            .delete();
+                                                      }
+                                                    } else {
+                                                      print(
+                                                          'No documents found with the matching image URL.');
+                                                    }
+                                                  });
+                                                });
                                               }
                                               await documentRef.delete();
                                               ScaffoldMessenger.of(context)
